commit_message_generation/pyproject.toml
[tool.poetry]
name = "baselines-cmg"
version = "0.1.1"
description = "Benchmark for Commit Message Generation task as part of BenchName."
authors = [
    "anon <anon@anon.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://anonymous.4open.science/r/icml-benchname-2025/README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = "^1.3.6"
hydra-core = "^1.3.2"
wandb = "^0.19.2"
jsonlines = "^3.1.0"
tiktoken = "^0.4.0"
torchmetrics = "^1.2.1"
evaluate = "^0.4.1"
bert-score = "^0.3.13"
sacrebleu = "^2.3.1"
rouge-score = "^0.1.2"
bitsandbytes = "^0.41.2"
accelerate = "^0.23.0"
optimum = "^1.14.1"
scipy = "^1.11.4"
backoff = "^2.2.1"
torch = "^2.5.1"
together = "^1.3.11"
python-dotenv = "^1.0.1"
langchain = "^0.3.16"
rank-bm25 = "^0.2.2"
gitpython = "^3.1.44"
langchain-community = "^0.3.16"
chardet = "^5.2.0"
datasets = "^3.2.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
mypy = "^1.5.0"
pytest = "^7.4.0"
ruff = "^0.9.1"

[tool.poetry.group.eda.dependencies]
jupyter = "^1.0.0"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
plotly = "^5.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "src/metrics/reused_implementations/b_norm.py"
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.isort]
profile = "black"
force_sort_within_sections = true
order_by_type = true

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
